+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Publish new state to `homeassistant/switch/switchbot/MAC_ADDRESS/state` on success
 
 ## [0.2.0] - 2020-05-08
 ### Added